fervent-api/requirements.txt
fastapi
pydantic
python-multipart
python-dotenv
unstructured
langchain
langchain_chroma
langchain_openai
langchain_community
pillow
pdf2image
pytesseract
spacy
pytest
pytest_asyncio